# src/Client/Logistics.DriverApp/composeApp/src/openapi/api-spec.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Logistics.API",
    "version": "1.0"
  },
  "paths": {
...
    "/loads/{id}": {
      "get": {
        "tags": [
          "Load"
        ],
        "operationId": "GetLoadById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Load"
        ],
        "operationId": "UpdateLoad",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoadCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoadCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoadCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Load"
        ],
        "operationId": "DeleteLoad",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/loads": {
      "get": {
        "tags": [
          "Load"
        ],
        "operationId": "GetLoads",
        "parameters": [
          {
            "name": "LoadAllPages",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "OnlyActiveLoads",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "UserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "TruckId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoPagedResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoPagedResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Load"
        ],
        "operationId": "CreateLoad",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoadCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoadCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoadCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "tags": [
          "Notification"
        ],
        "operationId": "GetNotifications",
        "parameters": [
          {
            "name": "StartDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "EndDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoArrayResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoArrayResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDtoArrayResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}": {
      "put": {
        "tags": [
          "Notification"
        ],
        "operationId": "UpdateNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "required": [
          "city",
          "country",
          "line1",
          "state",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "line1": {
            "type": "string",
            "nullable": true
          },
          "line2": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateLoadCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/LoadType"
          },
          "originAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "originLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "destinationAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "destinationLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "deliveryCost": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "assignedDispatcherId": {
            "type": "string",
            "format": "uuid"
          },
          "assignedTruckId": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EmployeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "lastKnownLocation": {
            "type": "string",
            "nullable": true
          },
          "lastKnownLocationLat": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lastKnownLocationLng": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "salaryType": {
            "$ref": "#/components/schemas/SalaryType"
          },
          "joinedDate": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GeoPoint": {
        "type": "object",
        "properties": {
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "$ref": "#/components/schemas/InvoiceType"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "stripeInvoiceId": {
            "type": "string",
            "nullable": true
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            },
            "nullable": true
          },
          "loadNumber": {
            "type": "integer",
            "format": "int64"
          },
          "loadId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "employee": {
            "$ref": "#/components/schemas/EmployeeDto"
          },
          "periodStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "subscriptionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "billingPeriodStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "billingPeriodEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceStatus": {
        "enum": [
          "draft",
          "issued",
          "partially_paid",
          "paid",
          "cancelled"
        ],
        "type": "string"
      },
      "InvoiceType": {
        "enum": [
          "load",
          "subscription",
          "payroll",
          "other"
        ],
        "type": "string"
      },
      "LoadDto": {
        "required": [
          "destinationAddress",
          "destinationLocation",
          "originAddress",
          "originLocation"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "originAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "originLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "destinationAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "destinationLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "deliveryCost": {
            "type": "number",
            "format": "double"
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "dispatchedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pickedUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "canConfirmPickUp": {
            "type": "boolean"
          },
          "canConfirmDelivery": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/LoadStatus"
          },
          "type": {
            "$ref": "#/components/schemas/LoadType"
          },
          "assignedDispatcherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assignedDispatcherName": {
            "type": "string",
            "nullable": true
          },
          "assignedTruckId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assignedTruckNumber": {
            "type": "string",
            "nullable": true
          },
          "currentAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "currentLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerDto"
          },
          "tripId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tripNumber": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "tripName": {
            "type": "string",
            "nullable": true
          },
          "assignedTruckDriversName": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoadDtoPagedResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadDto"
            },
            "nullable": true
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "LoadDtoResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "$ref": "#/components/schemas/LoadDto"
          }
        },
        "additionalProperties": false
      },
...
      "LoadStatus": {
        "enum": [
          "draft",
          "dispatched",
          "picked_up",
          "delivered",
          "cancelled"
        ],
        "type": "string"
      },
...
      "LoadType": {
        "enum": [
          "general_freight",
          "refrigerated_goods",
          "hazardous_materials",
          "oversize_heavy",
          "liquid",
          "bulk",
          "vehicle",
          "livestock"
        ],
        "type": "string"
      },
...
      "Money": {
        "required": [
          "amount",
          "currency"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "isRead": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "NotificationDtoArrayResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          },
          "methodId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PaymentStatus": {
        "enum": [
          "pending",
          "paid",
          "failed",
          "cancelled",
          "refunded",
          "partially_refunded"
        ],
        "type": "string"
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SalaryType": {
        "enum": [
          "none",
          "monthly",
          "weekly",
          "share_of_gross",
          "rate_per_mile",
          "rate_per_kilometer",
          "hourly"
        ],
        "type": "string"
      },
...
      "UpdateLoadCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/LoadType"
          },
          "originAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "originLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "destinationAddress": {
            "$ref": "#/components/schemas/Address"
          },
          "destinationLocation": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "deliveryCost": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "distance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "assignedDispatcherId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "assignedTruckId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "customerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/LoadStatus"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateNotificationCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isRead": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}